+++ crates/resvg/src/filter.rs
@@ -7,7 +7,7 @@ use std::rc::Rc;
 use rgb::FromSlice;
 use usvg::{FuzzyEq, FuzzyZero, Transform};
 
-use crate::geom::{ScreenRect, UsvgRectExt};
+use crate::geom::{IntRect, UsvgRectExt};
 use crate::tree::{ConvTransform, Node};
 
 // TODO: apply single primitive filters in-place
@@ -270,7 +270,7 @@ pub(crate) enum Error {
 
 trait PixmapExt: Sized {
     fn try_create(width: u32, height: u32) -> Result<tiny_skia::Pixmap, Error>;
-    fn copy_region(&self, region: ScreenRect) -> Result<tiny_skia::Pixmap, Error>;
+    fn copy_region(&self, region: IntRect) -> Result<tiny_skia::Pixmap, Error>;
     fn clear(&mut self);
     fn into_srgb(&mut self);
     fn into_linear_rgb(&mut self);
@@ -281,7 +281,7 @@ impl PixmapExt for tiny_skia::Pixmap {
         tiny_skia::Pixmap::new(width, height).ok_or(Error::InvalidRegion)
     }
 
-    fn copy_region(&self, region: ScreenRect) -> Result<tiny_skia::Pixmap, Error> {
+    fn copy_region(&self, region: IntRect) -> Result<tiny_skia::Pixmap, Error> {
         let rect =
             tiny_skia::IntRect::from_xywh(region.x(), region.y(), region.width(), region.height())
                 .ok_or(Error::InvalidRegion)?;
@@ -319,7 +319,7 @@ struct Image {
     /// Image's content outside this region will be transparent/cleared.
     ///
     /// Currently used only for `feTile`.
-    region: ScreenRect,
+    region: IntRect,
 
     /// The current color space.
     color_space: usvg::filter::ColorInterpolation,
@@ -330,7 +330,7 @@ impl Image {
         let (w, h) = (image.width(), image.height());
         Image {
             image: Rc::new(image),
-            region: ScreenRect::new(0, 0, w, h).unwrap(),
+            region: IntRect::new(0, 0, w, h).unwrap(),
             color_space,
         }
     }
@@ -392,7 +392,7 @@ struct FilterResult {
 
 pub fn apply(
     filter: &Filter,
-    region: ScreenRect,
+    region: IntRect,
     ts: &usvg::Transform,
     fill_paint: Option<&tiny_skia::Pixmap>,
     stroke_paint: Option<&tiny_skia::Pixmap>,
@@ -424,7 +424,7 @@ pub fn apply(
 fn apply_inner(
     filter: &Filter,
     inputs: &FilterInputs,
-    region: ScreenRect,
+    region: IntRect,
     ts: &usvg::Transform,
 ) -> Result<Image, Error> {
     let mut results: Vec<FilterResult> = Vec::new();
@@ -434,7 +434,7 @@ fn apply_inner(
         let mut subregion = primitive
             .rect
             .transform(ts)
-            .map(|r| r.to_screen_rect_round_out())
+            .map(|r| r.to_int_rect_round_out())
             .ok_or(Error::InvalidRegion)?;
 
         // `feOffset` inherits its region from the input.
@@ -688,11 +688,11 @@ fn calc_subregion(
 
 fn get_input(
     input: &usvg::filter::Input,
-    region: ScreenRect,
+    region: IntRect,
     inputs: &FilterInputs,
     results: &[FilterResult],
 ) -> Result<Image, Error> {
-    let convert = |in_image: Option<&tiny_skia::Pixmap>, region: ScreenRect| {
+    let convert = |in_image: Option<&tiny_skia::Pixmap>, region: IntRect| {
         let image = if let Some(image) = in_image {
             image.clone()
         } else {
@@ -868,7 +868,7 @@ fn apply_offset(
 fn apply_blend(
     fe: &usvg::filter::Blend,
     cs: usvg::filter::ColorInterpolation,
-    region: ScreenRect,
+    region: IntRect,
     input1: Image,
     input2: Image,
 ) -> Result<Image, Error> {
@@ -904,7 +904,7 @@ fn apply_blend(
 fn apply_composite(
     fe: &usvg::filter::Composite,
     cs: usvg::filter::ColorInterpolation,
-    region: ScreenRect,
+    region: IntRect,
     input1: Image,
     input2: Image,
 ) -> Result<Image, Error> {
@@ -968,7 +968,7 @@ fn apply_composite(
 fn apply_merge(
     fe: &usvg::filter::Merge,
     cs: usvg::filter::ColorInterpolation,
-    region: ScreenRect,
+    region: IntRect,
     inputs: &FilterInputs,
     results: &[FilterResult],
 ) -> Result<Image, Error> {
@@ -990,7 +990,7 @@ fn apply_merge(
     Ok(Image::from_image(pixmap, cs))
 }
 
-fn apply_flood(fe: &usvg::filter::Flood, region: ScreenRect) -> Result<Image, Error> {
+fn apply_flood(fe: &usvg::filter::Flood, region: IntRect) -> Result<Image, Error> {
     let c = fe.color;
 
     let mut pixmap = tiny_skia::Pixmap::try_create(region.width(), region.height())?;
@@ -1007,7 +1007,7 @@ fn apply_flood(fe: &usvg::filter::Flood, region: ScreenRect) -> Result<Image, Er
     ))
 }
 
-fn apply_tile(input: Image, region: ScreenRect) -> Result<Image, Error> {
+fn apply_tile(input: Image, region: IntRect) -> Result<Image, Error> {
     let subregion = input.region.translate(-region.x(), -region.y());
 
     let tile_pixmap = input.image.copy_region(subregion)?;
@@ -1033,8 +1033,8 @@ fn apply_tile(input: Image, region: ScreenRect) -> Result<Image, Error> {
 
 fn apply_image(
     fe: &usvg::filter::Image,
-    region: ScreenRect,
-    subregion: ScreenRect,
+    region: IntRect,
+    subregion: IntRect,
     ts: &usvg::Transform,
 ) -> Result<Image, Error> {
     let mut pixmap = tiny_skia::Pixmap::try_create(region.width(), region.height())?;
@@ -1200,7 +1200,7 @@ fn apply_morphology(
 
 fn apply_displacement_map(
     fe: &usvg::filter::DisplacementMap,
-    region: ScreenRect,
+    region: IntRect,
     cs: usvg::filter::ColorInterpolation,
     ts: &usvg::Transform,
     input1: Image,
@@ -1228,7 +1228,7 @@ fn apply_displacement_map(
 
 fn apply_turbulence(
     fe: &usvg::filter::Turbulence,
-    region: ScreenRect,
+    region: IntRect,
     cs: usvg::filter::ColorInterpolation,
     ts: &usvg::Transform,
 ) -> Result<Image, Error> {
@@ -1260,7 +1260,7 @@ fn apply_turbulence(
 
 fn apply_diffuse_lighting(
     fe: &usvg::filter::DiffuseLighting,
-    region: ScreenRect,
+    region: IntRect,
     cs: usvg::filter::ColorInterpolation,
     ts: &usvg::Transform,
     input: Image,
@@ -1283,7 +1283,7 @@ fn apply_diffuse_lighting(
 
 fn apply_specular_lighting(
     fe: &usvg::filter::SpecularLighting,
-    region: ScreenRect,
+    region: IntRect,
     cs: usvg::filter::ColorInterpolation,
     ts: &usvg::Transform,
     input: Image,
@@ -1307,7 +1307,7 @@ fn apply_specular_lighting(
 
 fn transform_light_source(
     mut source: usvg::filter::LightSource,
-    region: ScreenRect,
+    region: IntRect,
     ts: &Transform,
 ) -> usvg::filter::LightSource {
     use std::f64::consts::SQRT_2;