+++ crates/resvg/src/lib.rs
@@ -28,5 +28,5 @@ mod path;
 mod render;
 mod tree;
 
-pub use crate::geom::{ScreenRect, ScreenSize};
+pub use crate::geom::{IntRect, IntSize};
 pub use crate::tree::Tree;