+++ crates/resvg/src/main.rs
@@ -415,17 +415,17 @@ enum FitTo {
 }
 
 impl FitTo {
-    fn fit_to_size(&self, size: resvg::ScreenSize) -> Option<resvg::ScreenSize> {
+    fn fit_to_size(&self, size: resvg::IntSize) -> Option<resvg::IntSize> {
         match *self {
             FitTo::Original => Some(size),
             FitTo::Width(w) => size.scale_to_width(w),
             FitTo::Height(h) => size.scale_to_height(h),
-            FitTo::Size(w, h) => resvg::ScreenSize::new(w, h).map(|s| size.scale_to(s)),
+            FitTo::Size(w, h) => resvg::IntSize::new(w, h).map(|s| size.scale_to(s)),
             FitTo::Zoom(z) => size.scale_by(z as f64),
         }
     }
 
-    fn fit_to_transform(&self, size: resvg::ScreenSize) -> tiny_skia::Transform {
+    fn fit_to_transform(&self, size: resvg::IntSize) -> tiny_skia::Transform {
         let size1 = size.to_size();
         let size2 = match self.fit_to_size(size) {
             Some(v) => v.to_size(),
@@ -649,7 +649,7 @@ fn render_svg(args: &Args, tree: &usvg::Tree) -> Result<tiny_skia::Pixmap, Strin
 
         let size = args
             .fit_to
-            .fit_to_size(resvg::ScreenSize::from_usvg(bbox.size()))
+            .fit_to_size(resvg::IntSize::from_usvg(bbox.size()))
             .ok_or_else(|| "target size is zero".to_string())?;
 
         // Unwrap is safe, because `size` is already valid.
@@ -663,7 +663,7 @@ fn render_svg(args: &Args, tree: &usvg::Tree) -> Result<tiny_skia::Pixmap, Strin
 
         let ts = args
             .fit_to
-            .fit_to_transform(resvg::ScreenSize::from_usvg(tree.size));
+            .fit_to_transform(resvg::IntSize::from_usvg(tree.size));
 
         let rtree = resvg::Tree::from_usvg_node(&node)
             .ok_or_else(|| "zero-size node detected".to_string())?;
@@ -675,7 +675,7 @@ fn render_svg(args: &Args, tree: &usvg::Tree) -> Result<tiny_skia::Pixmap, Strin
 
             let size = args
                 .fit_to
-                .fit_to_size(resvg::ScreenSize::from_usvg(tree.size))
+                .fit_to_size(resvg::IntSize::from_usvg(tree.size))
                 .ok_or_else(|| "target size is zero".to_string())?;
 
             // Unwrap is safe, because `size` is already valid.
@@ -700,7 +700,7 @@ fn render_svg(args: &Args, tree: &usvg::Tree) -> Result<tiny_skia::Pixmap, Strin
     } else {
         let size = args
             .fit_to
-            .fit_to_size(resvg::ScreenSize::from_usvg(tree.size))
+            .fit_to_size(resvg::IntSize::from_usvg(tree.size))
             .ok_or_else(|| "target size is zero".to_string())?;
 
         // Unwrap is safe, because `size` is already valid.
@@ -712,7 +712,7 @@ fn render_svg(args: &Args, tree: &usvg::Tree) -> Result<tiny_skia::Pixmap, Strin
 
         let ts = args
             .fit_to
-            .fit_to_transform(resvg::ScreenSize::from_usvg(tree.size));
+            .fit_to_transform(resvg::IntSize::from_usvg(tree.size));
 
         let rtree = resvg::Tree::from_usvg(tree);
         rtree.render(ts, &mut pixmap.as_mut());
@@ -749,9 +749,9 @@ fn trim_pixmap(
         transform.ty as f64,
     );
 
-    let limit = resvg::ScreenRect::new(0, 0, pixmap.width(), pixmap.height()).unwrap();
+    let limit = resvg::IntRect::new(0, 0, pixmap.width(), pixmap.height()).unwrap();
 
-    let content_area = resvg::ScreenRect::from_usvg(content_area.transform(&uts)?);
+    let content_area = resvg::IntRect::from_usvg(content_area.transform(&uts)?);
     let content_area = content_area.fit_to_rect(limit);
     let content_area = tiny_skia::IntRect::from_xywh(
         content_area.x(),