+++ CHANGELOG.md
@@ -35,14 +35,16 @@ This changelog also contains important changes in dependencies.
 - (c-api) Removed `fit_to` argument from `resvg_render_node`.
 - `usvg::ScreenSize` moved to `resvg`.
 - `usvg::ScreenRect` moved to `resvg`.
+- Rename `resvg::ScreenSize` into `resvg::IntSize`.
+- Rename `resvg::ScreenRect` into `resvg::IntRect`.
 
 ### Removed
 - `filter` build feature from `resvg`. Filters are always enabled now.
 - `resvg::FitTo`
 - `usvg::utils::view_box_to_transform_with_clip`
-- `usvg::Size::to_screen_size`. Use `resvg::ScreenSize::from_usvg` instead.
-- `usvg::Rect::to_screen_size`. Use `resvg::ScreenSize::from_usvg(rect.size())` instead.
-- `usvg::Rect::to_screen_rect`. Use `resvg::ScreenRect::from_usvg` instead.
+- `usvg::Size::to_screen_size`. Use `resvg::IntSize::from_usvg` instead.
+- `usvg::Rect::to_screen_size`. Use `resvg::IntSize::from_usvg(rect.size())` instead.
+- `usvg::Rect::to_screen_rect`. Use `resvg::IntRect::from_usvg` instead.
 - (c-api) `resvg_fit_to`
 - (c-api) `resvg_fit_to_type`